App/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.13
django-allauth==0.51.0
django-environ==0.8.1
django-heroku==0.3.1
django-sslserver==0.22
django-storages==1.12.3
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.7
python3-openid==3.2.0
pytz==2021.3
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.14
whitenoise==5.3.0
Pillow==9.4.0